0812_m1 #Mypage Route 연결 및 javascript 설정 중
